refactor app and index.js
